% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_icd_category.R
\name{add_icd_category}
\alias{add_icd_category}
\title{Add a column of ICD categories to a data.frame}
\usage{
add_icd_category(icd, icd_column, icd_type = c("diagnosis", "procedure"),
  icd_lookup_column = "global_short_description", parallel = FALSE,
  cluster_num = parallel::detectCores() - 1)
}
\arguments{
\item{icd}{A data.frame with ICD 9 or 10 diagnoses.}

\item{icd_column}{The name of the column of ICD codes.}

\item{icd_type}{Are these disease or diagnosis codes.}

\item{icd_lookup_column}{Which column from \code{icd_lookup} should be returned. You
can see this data frame by running \code{data(icd_lookup)}.}

\item{parallel}{Use parallel processing?}

\item{cluster_num}{Number of clusters to use if \code{parallel} is TRUE.}
}
\description{
Add a column of ICD categories to a data.frame
}
