% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_groups.R
\name{corr_groups}
\alias{corr_groups}
\title{Find highly correlated groups for dimension reduction}
\usage{
corr_groups(data, cutoff = 0.9, cor_mat = FALSE)
}
\arguments{
\item{data}{A data.frame or matrix of features for modeling}

\item{cutoff}{Minimum correlation used to make groups}

\item{cor_mat}{If TRUE, a list of correlation matrices are returned}
}
\description{
Find highly correlated groups for dimension reduction
}
\examples{
corr_groups(mtcars)
corr_groups(mtcars, cor_mat = TRUE)
}
