% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_corr_groups.R
\name{find_corr_groups}
\alias{find_corr_groups}
\title{Find highly correlated groups for dimension reduction}
\usage{
find_corr_groups(data, cutoff = 0.8, cor_mat = FALSE)
}
\arguments{
\item{data}{A data.frame or matrix of features for modeling}

\item{cutoff}{Minimum correlation used to make groups}

\item{cor_mat}{If TRUE, a list of correlation matrices are returned}
}
\description{
Find highly correlated groups for dimension reduction
}
\examples{
find_corr_groups(mtcars)
find_corr_groups(mtcars, cor_mat = TRUE)
}
