% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lab_data.R
\name{generate_lab_data}
\alias{generate_lab_data}
\title{Generate data to test lab budget project}
\usage{
generate_lab_data(years, by = c("day", "week", "month", "year"),
  mean_volume = 100, sd_volume = 20, trend = FALSE, seasonality = FALSE,
  predictor_variables = 0, noisy_variables = 0)
}
\arguments{
\item{years}{How many years of data}

\item{by}{What time duration do you want}

\item{mean_volume}{What is the mean volume for this lab test}

\item{sd_volume}{What is the standard deviation for these tests}

\item{trend}{Do you want a trend over time}

\item{seasonality}{Do you want seasonality during the year}

\item{predictor_variables}{How many predictor variables do you want}

\item{noisy_variables}{How many non-predictive variables do you want}
}
\description{
Generate data to test lab budget project
}
\examples{
library(dplyr)
library(tidyr)
library(ggplot2)

x <- generate_lab_data(years = 3, by = "week", mean_volume = 100, 
                       sd_volume = 20, trend = TRUE, seasonality = TRUE, 
                       predictor_variables = 2, noisy_variables = 1) 

lm(test_volume ~ noisy1, data = x) \%>\% summary()
lm(test_volume ~ predictor1, data = x) \%>\% summary()
lm(test_volume ~ predictor2, data = x) \%>\% summary()
lm(test_volume ~ predictor1 + predictor2, data = x) \%>\% summary()

x <- tidyr::gather(x, key = "key", value = "value", test_volume:noisy1)

ggplot(x, aes(index, value, color = key)) + geom_point()
}
