% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_MRNs.R
\name{match_MRNs}
\alias{match_MRNs}
\title{Match MRNs from two EMRs}
\usage{
match_MRNs(x, y)
}
\arguments{
\item{x, y}{Two data frames with `MRN`, `first_name`, `last_name`, and 
`birth_date` columns}
}
\description{
Match MRNs from two EMRs
}
