% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_sankey.R
\name{process_sankey}
\alias{process_sankey}
\title{Process a data.frame for a Sankey network graph}
\usage{
process_sankey(dat, lineage, value, drop_NAs = TRUE)
}
\arguments{
\item{dat}{A \code{data.frame}.}

\item{lineage}{A vector of column names in descending order.}

\item{value}{Name of the value column.}

\item{drop_NAs}{Logical indicating what should be done with NA sources and targets.}
}
\value{
A list of two \code{data.frame}s: links and nodes.
}
\description{
Process a data.frame for a Sankey network graph
}
\examples{
library(EDW)
library(networkD3)

edw <- connect_edw() \%>\%
 tbl_(c(e = "facilitybillline"), schema = "finance") \%>\%
 inner_join_(c(en = "encounter"), c("e.[encounter id]" = "en.[encounter id]"),
             schema = "population") \%>\%
 filter_("e.[revenue group category] == 'LAB'") \%>\%
 select_(c("en.[encounter type]", "en.[GL company]", "e.[charge line total]")) \%>\%
 group_by_(c("en.[encounter type]", "en.[GL company]")) \%>\%
 summarise_(c("charge_total" = "sum(e.[charge line total])")) \%>\%
 run_query()

edw <- process_sankey(dat = edw, c("encounter type", "GL company"),
                     value = "charge_total")

sankeyNetwork(Links = edw$links, Nodes = edw$nodes, Source = "source",
             Target = "target", Value = "value", NodeID = "name",
             units = "dollas")
}
