% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_iuh.R
\name{iuh_pal}
\alias{iuh_pal}
\alias{scale_color_iuh}
\alias{scale_colour_iuh}
\alias{scale_fill_iuh}
\title{IU Health color palette (discrete)}
\usage{
iuh_pal(fill = TRUE)

scale_color_iuh(...)

scale_colour_iuh(...)

scale_fill_iuh(...)
}
\arguments{
\item{fill}{Use the fill palette.}
}
\description{
IU Health color palette (discrete)

IU Health color scales

IU Health color fill scales
}
