% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model_features.R
\name{select_model_features}
\alias{select_model_features}
\title{Select features to use in a model}
\usage{
select_model_features(features, corr_cutoff = 0.9, rank_table)
}
\arguments{
\item{features}{A data frame of numeric features}

\item{corr_cutoff}{Minimum correlation used to make groups}

\item{rank_table}{A data frame of information for selecting features. The
first column should be named "variable" and have the name of each column in
the feature data frame. Each column after "variable" should be numeric column
for ranking the features. The order for each column should be increasing, i.e.
a rank of 1 will beat a rank of 2.}
}
\description{
Select features to use in a model
}
\details{
The point to this function is to eliminate features in your data frame
based on the information given to the function. This information can be
expertly generated, automatically generated, or both. The fundamental assumption
is that you will want to remove highly correlated information. The rank_table
supplies the information needed to choose the best member of each highly
correlated group. The features data frame is returned with the chosen features.
The columns will be ordered from highest ranked (left) to lowest rank,
along with an attribute giving the rank information. Only numeric features will
be considered.
}
