% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_iuh.R
\name{theme_iuh}
\alias{theme_iuh}
\title{ggplot2 theme for IU Health graphs}
\usage{
theme_iuh(dark = FALSE)
}
\description{
ggplot2 theme for IU Health graphs
}
\examples{
library(ggplot2)

ggplot(mpg, aes(class, fill = drv)) + geom_bar() + theme_iuh() + 
  scale_fill_iuh() + ggtitle("IU Health")
}
