% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll.R
\name{unroll}
\alias{unroll}
\title{Recover original data series from moving average}
\usage{
unroll(
  moving_series,
  original_seed,
  window_length,
  align = c("left", "right"),
  window_function = c("mean", "sum")
)
}
\arguments{
\item{moving_series}{a vector of moving averages}

\item{original_seed}{a vector of the first or last numbers of the original series--must be \code{window_length - 1} or longer}

\item{window_length}{integer length of the window}

\item{align}{character specifying whether the index of the original series was left- or right-aligned or centered (default) compared to the rolling window of series}

\item{window_function}{what function was used to roll up the original data}
}
\description{
Recover original data series from moving average
}
