% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unroll.R
\name{unroll_year_to_date}
\alias{unroll_year_to_date}
\title{Recover original series from a vector of year-to-date averages}
\usage{
unroll_year_to_date(x)
}
\arguments{
\item{x}{a vector of year-to-date averages}
}
\description{
Recover original series from a vector of year-to-date averages
}
\examples{
library(dplyr)

set.seed(1)
dat <- dplyr::tibble(year = rep(2018:2019, each = 12),
                     month = rep(1:12, times = 2),
                     original = sample(200:300, 24, TRUE)) \%>\%
  dplyr::group_by(year) \%>\%
  dplyr::mutate(ytd = roll_mean_year_to_date(original),
                unrolled = unroll_year_to_date(ytd))
}
