% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_for_headers.R
\name{cleanup_for_headers}
\alias{cleanup_for_headers}
\title{Transform vector into strings that can be headers in a data.frame}
\usage{
cleanup_for_headers(x)
}
\arguments{
\item{x}{Character vector of column names}
}
\description{
Transform vector into strings that can be headers in a data.frame
}
\examples{
column_names <- c("MRN ID", "First & Middle Name", "Name: Last")

cleanup_for_headers(column_names)
}
