% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_MRNs.R
\name{match_MRNs}
\alias{match_MRNs}
\title{Match MRNs from two EMRs}
\usage{
match_MRNs(x, y)
}
\arguments{
\item{x, y}{Two data frames with columns `mrn`, `first_name`, `last_name`, and 
`birth_date` (yyyy-mm-dd)}
}
\value{
A data.frame with column names with suffixes '_x' and '_y' corresponding
to the x and y inputs. The dist column gives the string distance between the 
two first names.
}
\description{
Takes two data.frame objects with patient information and matches them
}
\details{
The function cleans up the last names slightly and then matches exactly
on the cleaned up last names and the birth dates. Then fuzzy matching is done
on the first name, using the \code{stringdist} package.
}
\examples{
data(y_mrn)
data(x_mrn)

match_MRNs(x_mrn, y_mrn)
}
